chart 모듈화
