+++ components/Stats/Chart/index.ts
@@ -1,0 +1,7 @@
+import { Line } from "./Line";
+import { Pie } from "./Pie";
+
+export {
+  Line,
+  Pie
+}
